deploy fixes

+++ config/deploy.rb
@@ -34,6 +34,17 @@ set :keep_releases, 5
 # 
 # before 'deploy:finalize_update', 'deploy:symlink_db'
 # 
+#
+namespace :puma do
+  task :start, roles: :app do
+    puts 'Nothing to do here, because runit running...'
+  end
+
+  task :restart, roles: :app do
+    puts 'Nothing to do here, because runit running...'
+  end
+end
+
 namespace :deploy do
   desc "Symlinks the database.yml"
   task :symlink_db, roles: :app do
@@ -45,15 +56,27 @@ namespace :deploy do
     run "cd #{current_path} && RAILS_ENV=#{rails_env} #{rake} db:seed"
   end
 
+
   namespace :assets do
     desc "Local precompile assets and upload to server"
     task :precompile, roles: :app do
-      run_locally "RAILS_ENV=#{rails_env} #{rake} assets:clean && RAILS_ENV=#{rails_env} #{rake} assets:precompile"
-      run_locally "cd public && tar -jcf assets.tar.bz2 assets"
-      top.upload "public/assets.tar.bz2", "#{shared_path}", via: :scp
-      run "cd #{shared_path} && tar -jxf assets.tar.bz2 && rm assets.tar.bz2"
-      run_locally "rm public/assets.tar.bz2"
-      run_locally "#{rake} assets:clean"
+      from = source.next_revision(current_revision)
+      if capture("cd #{latest_release} && #{source.local.log(from)} vendor/assets/ lib/assets/ app/assets/ | wc -l").to_i > 0
+        run_locally "RAILS_ENV=#{rails_env} #{rake} assets:clean && RAILS_ENV=#{rails_env} #{rake} assets:precompile"
+        run_locally "cd public && tar -jcf assets.tar.bz2 assets"
+        top.upload "public/assets.tar.bz2", "#{shared_path}", via: :scp
+        run "cd #{shared_path} && tar -jxf assets.tar.bz2 && rm assets.tar.bz2"
+        run_locally "rm public/assets.tar.bz2"
+        run_locally "#{rake} assets:clean"
+      else
+        logger.info "Skipping asset precompilation because there were no asset changes"
+      end
+    #   run_locally "RAILS_ENV=#{rails_env} #{rake} assets:clean && RAILS_ENV=#{rails_env} #{rake} assets:precompile"
+    #   run_locally "cd public && tar -jcf assets.tar.bz2 assets"
+    #   top.upload "public/assets.tar.bz2", "#{shared_path}", via: :scp
+    #   run "cd #{shared_path} && tar -jxf assets.tar.bz2 && rm assets.tar.bz2"
+    #   run_locally "rm public/assets.tar.bz2"
+    #   run_locally "#{rake} assets:clean"
     end
 
     desc "Symlink local precompile assets"
@@ -78,4 +101,4 @@ before 'deploy:finalize_update', 'deploy:symlink_db'
 before 'deploy:finalize_update', 'deploy:assets:symlink'
 after 'deploy:update_code', 'deploy:assets:precompile'
 after "deploy:update", "deploy:cleanup"
-after 'deploy:restart', 'puma:stop'
+# after 'deploy:restart', 'puma:stop'
